% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{compute.rad}
\alias{compute.rad}
\title{Run test for unjustified disparate impact}
\usage{
compute.rad(r, controls = NULL)
}
\arguments{
\item{r}{object of class rad ("undi")}

\item{controls}{character vector of additional controls to consider in the
second-stage model}
}
\value{
tidy data frame of rad coefficients
}
\description{
Run test for unjustified disparate impact
}
