% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compute_bm}
\alias{compute_bm}
\title{Compute benchmark test for disparate impact of a policy}
\usage{
compute_bm(pol, controls = NULL, kitchen_sink = FALSE, base_group = NULL,
  minority_groups = NULL)
}
\arguments{
\item{pol}{object of class policy}

\item{controls}{character vector of additional controls to consider (i.e.,
valid benchmarks)}

\item{kitchen_sink}{logical; if TRUE, ignore \code{controls} argument, and
include all variables given with the policy, i.e., policy$features
(default: FALSE)}

\item{base_group}{(Optional) single group that acts as the pivot/base; by
default, if the grouping variable is a factor, set to the first level,
otherwise set to the first of sorted unique values}

\item{minority_groups}{(Optional) groups to compare to the base group; by
default, set to every unique value other than the base group}
}
\value{
tidy data frame of benchmark coefficients
}
\description{
Compute benchmark test for disparate impact of a policy
}
