% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compute_ot}
\alias{compute_ot}
\title{Compute outcome tests for disparate impact of a policy}
\usage{
compute_ot(pol, controls = NULL)
}
\arguments{
\item{pol}{object of class policy}

\item{controls}{character vector of additional controls to consider (i.e.,
conditional groupings)}
}
\value{
data frame, grouped by the group and additional columns specified in
  the \code{controls} arguments and corresponding \code{hitrate}, i.e.,
  \code{P(outcome = 1 | treatment = risk_treatment)}, where
  \code{risk_treatment = ifelse(pol$risk_col == "resp_trt", 1, 0)}.
}
\description{
Compute outcome tests for disparate impact of a policy
}
