% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compute_rad}
\alias{compute_rad}
\title{Compute risk-adjusted disparate impact estimate for a policy}
\usage{
compute_rad(pol, controls = NULL, base_group = NULL,
  minority_groups = NULL, fit_fn = "logit_coef", down_sample = 1,
  seed = round(stats::runif(1) * 10000), ...)
}
\arguments{
\item{pol}{object of class policy}

\item{controls}{character vector of additional controls to consider in the
second-stage model}

\item{base_group}{(Optional) single group that acts as the pivot/base; by
default, if the grouping variable is a factor, set to the first level,
otherwise set to the first of sorted unique values}

\item{minority_groups}{(Optional) groups to compare to the base group; by
default, set to every unique value other than the base group}

\item{fit_fn}{string indicating the rad estimation model/procedure used.
\code{*_coef} methods use models without interaction between risk and
group, and return the coeficient on group membership. \code{*_avg} methods
will fit more flexible models (possibly with interactions), and compute
average ratios across the population. (TODO: better documentation is
expected)}

\item{down_sample}{(Optional) proportion (between 0 and 1) or number (greater
than 1) of rows to sample, if down sampling the (test) data; default is 1
(i.e., use all data)}

\item{seed}{random seed to set}

\item{...}{additional arguments to pass to \code{fit} function from
\code{\link{rad_control}} for fine-tuning}
}
\value{
tidy data frame with columns \item{term}{the group members considered
  minority} \item{estimate}{log-odds of treatment, relative to base_group
  (equivalent to logistic regression coefficient)}
  \item{std.error}{coefficient standard errors for \code{*_coef} methods; TO
  BE IMPLEMENTED for \code{*_avg} methods (FOR NOW, ALL ZERO)!}
  \item{statistic/p.value}{(for \code{*_coef} methods) corresponding values
  from model fit} \item{ptrt_base/minor}{(for \code{*_avg} methods) estimated
  average treatment probability for base/minority groups} \item{method}{the
  method used} \item{controls}{features controlled for}
}
\description{
Compute risk-adjusted disparate impact estimate for a policy
}
