% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compute_rad_old}
\alias{compute_rad_old}
\title{Compute risk-adjusted disparate impact estimate for a policy}
\usage{
compute_rad_old(pol, controls = NULL, base_group = NULL,
  minority_groups = NULL, difun = NULL, down_sample = 1,
  seed = round(stats::runif(1) * 10000))
}
\arguments{
\item{pol}{object of class policy}

\item{controls}{character vector of additional controls to consider in the
second-stage model}

\item{base_group}{(Optional) single group that acts as the pivot/base; by
default, if the grouping variable is a factor, set to the first level,
otherwise set to the first of sorted unique values}

\item{minority_groups}{(Optional) groups to compare to the base group; by
default, set to every unique value other than the base group}

\item{difun}{a "disparate impact function" of the form \code{f(data, weights
= NULL)} used for predicting treatment conditional on group membership;
using \code{glm} with \code{family = quasibinomial} by default; the
\code{weights} argument is only used for sensitivity and *must* be
initialized to \code{NULL} (or the equivalent of non-weighted fitting)}

\item{down_sample}{(Optional) proportion (between 0 and 1) or number (greater
than 1) of rows to sample, if down sampling the (test) data; default is 1
(i.e., use all data)}

\item{seed}{random seed to set}
}
\value{
tidy data frame of rad coefficients
}
\description{
Compute risk-adjusted disparate impact estimate for a policy
}
