% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{di_models}
\alias{di_models}
\title{Get list of formula and functions to fit and predict treatment given risk,
group, and controls}
\usage{
di_models(pol, controls = NULL, fit_fn = c("logit"), ...)
}
\arguments{
\item{pol}{a \code{\link{policy}} object}

\item{controls}{character vector of additional controls to consider in the
second-stage model}

\item{fit_fn}{string indicating the fitting proceedure used.}

\item{...}{other arguments passed to the \code{fit} function}
}
\value{
a list with model types (e.g., glm/gbm), each with the original
  \code{formula}, and appropriate \code{$fit} and \code{$pred} functions
}
\description{
Get list of formula and functions to fit and predict treatment given risk,
group, and controls
}
