% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr.policy}
\alias{dplyr.policy}
\alias{filter.policy}
\alias{mutate.policy}
\alias{group_by.policy}
\alias{ungroup.policy}
\title{Generate a new policy after applying various dplyr verbs}
\usage{
\method{filter}{policy}(.data, ...)

\method{mutate}{policy}(.data, ...)

\method{group_by}{policy}(.data, ...)

\method{ungroup}{policy}(x, ...)
}
\arguments{
\item{.data}{a policy object}

\item{...}{arguments passed to corresponding dplyr verb}

\item{x}{a policy object}
}
\value{
new policy object with altered data
}
\description{
Generate a new policy after applying various dplyr verbs
}
\details{
Note that the policy object is \emph{never} refit, and only the
  \code{$data} element is manipulated. In certain cases (e.g., after a
  filter), the policy might need to be refit via
  \code{\link{estimate_policy}}.
}
