% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{estimate_policy}
\alias{estimate_policy}
\title{(Re)Estimate treatment and response surface models for a policy}
\usage{
estimate_policy(pol, features = NULL, ...)
}
\arguments{
\item{pol}{a policy object}

\item{features}{(Optional) character vector of features to use; if
\code{NULL}, \code{pol$features} is used}

\item{...}{additional arguments  passed to model fitting functions \code{fit1}
and \code{fit_ptreat}. See \code{\link{policy}}}
}
\value{
object of class \code{policy} where the relative columns in
  \code{$data} are updated with new predictions
}
\description{
(Re)Estimate treatment and response surface models for a policy
}
