% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Wrapper for fitting \code{\link[glmnet]{glmnet}} models from a formula}
\usage{
fit_glmnet(f, d, ...)
}
\arguments{
\item{f}{formula}

\item{d}{data object}

\item{...}{additional arguments for \code{\link[glmnet]{cv.glmnet}}}
}
\value{
an \code{\link[glmnet]{cv.glmnet}} object
}
\description{
Wrapper for fitting \code{\link[glmnet]{glmnet}} models from a formula
}
