% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{List of paired fit/pred methods by type (glm/glmnet/gbm)}
\usage{
models()
}
\value{
a list with model types (e.g., glm/gbm), each with appropriate
  \code{$fit} and \code{$pred} functions
}
\description{
List of paired fit/pred methods by type (glm/glmnet/gbm)
}
