% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optimsens}
\alias{optimsens}
\title{Find parameter values that min/maximize sensitivity results}
\usage{
optimsens(pol, range_q = c(0, 1), range_dp = c(0, log(2)), range_d0 = c(0,
  log(2)), range_d1 = c(0, log(2)), base_group = NULL,
  minority_groups = NULL, range_q_ratio = NULL, controls = NULL,
  debug = FALSE)
}
\arguments{
\item{pol}{object of class \code{policy}}

\item{range_q}{2D vector specifying min/max value of p(u = 1 | x)}

\item{range_dp}{2D vector specifying min/max value of change in log-odds of
treat = 1 if u = 1}

\item{range_d0}{2D vector specifying min/max value of change in log-odds of
response = 1 if treat = 0 and u = 1}

\item{range_d1}{2D vector specifying min/max value of change in log-odds of
response = 1 of treat = 1 and u = 1}

\item{base_group}{(Optional) single group that acts as the pivot/base; by
default, if the grouping variable is a factor, set to the first level,
otherwise set to the first of sorted unique values}

\item{minority_groups}{(Optional) groups to compare to the base group; by
default, set to every unique value other than the base group}

\item{range_q_ratio}{(Optional) 2D vector. If set, the minority and majority
values of q will not be allowed to vary independently, but instead will be
constrained to vary by the given range of log odds. ie q_minority =
inv.logit(logit(q_majority) + u), where u is in range_q_ratio}

\item{controls}{vector of legitimate controls to use; the ones specified
within the policy object will be used if not specified}

\item{debug}{logical flag, if TRUE, returns a list of results and the
expanded data frame used to fit model}
}
\value{
a list with the following elements \item{results}{\code{tidy}
  dataframe of second-stage model coefficients after searching for min/max
  values across specified sensitivy parameter ranges, independently for each
  minority group} \item{optim}{nested data frame where the \code{$optim}
  column contains the optimization results}
}
\description{
Within specified range of sensitivity parameters, find the ones that achieve
minimum/maximum sensitivity results
}
