% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.optimsens}
\alias{plot.optimsens}
\title{Plot optimized sensitivity results for a policy}
\usage{
\method{plot}{optimsens}(x, ...)
}
\arguments{
\item{x}{object of class \code{optimsens}}

\item{...}{ignored; included for S3 generic/method consistency}
}
\value{
a ggplot object
}
\description{
Plot optimized sensitivity results for a policy
}
