% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.policy}
\alias{plot.policy}
\title{Plot min/max sensitivity results for a policy from a \code{sens} object}
\usage{
\method{plot}{policy}(x, nbins = 10, ...)
}
\arguments{
\item{x}{object of class \code{policy}}

\item{nbins}{number of bins to use (e.g., for calibration plots)}

\item{...}{ignored; included for S3 generic/method consistency}
}
\value{
a named list of ggplot object
}
\description{
Plot min/max sensitivity results for a policy from a \code{sens} object
}
