% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sensitive_policy}
\alias{plot.sensitive_policy}
\title{Plot changes in risk-treatment for a \code{sensitive_policy}}
\usage{
\method{plot}{sensitive_policy}(x, down_sample = 30, method = "glm",
  method.args = list(family = "quasibinomial"), ...)
}
\arguments{
\item{x}{object of class \code{sensitive_policy}}

\item{down_sample}{(Optional) proportion (between 0 and 1) or number (greater
than 1) of rows to sample from each group, if down sampling the data;
default is 30}

\item{method}{passed to \code{\link[ggplot2]{geom_smooth}}}

\item{method.args}{passed to \code{\link[ggplot2]{geom_smooth}}}

\item{...}{ignored; included for S3 generic/method consistency}
}
\value{
a ggplot object
}
\description{
Plot changes in risk-treatment for a \code{sensitive_policy}
}
