% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict4mm}
\alias{predict4mm}
\title{Wrapper to compute predictions from a model matrix}
\usage{
predict4mm(m, d, f, ...)
}
\arguments{
\item{m}{model object}

\item{d}{data frame to generate predictions on}

\item{f}{model formula; used to transform \code{d} to a \code{design matrix}}

\item{...}{other arguments to pass to the predict function}
}
\value{
generated predictions
}
\description{
Wrapper to compute predictions from a model matrix
}
