% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{predict_sgd}
\alias{predict_sgd}
\title{Wrapper to compute predictions for sgd model}
\usage{
predict_sgd(m, d, f, type = "response", ...)
}
\arguments{
\item{m}{sgd object}

\item{d}{data frame to generate predictions on}

\item{f}{formula used in sgd object; used to transform \code{d} to a
\code{design matrix}}

\item{type}{the type of prediction to return (default: response)}

\item{...}{other arguments to pass to predict.sgd}
}
\value{
prediction for sgd model
}
\description{
Wrapper to compute predictions for sgd model
}
