% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.optimsens}
\alias{print.optimsens}
\title{\code{print} method of object class \code{\link{optimsens}}}
\usage{
\method{print}{optimsens}(x, ...)
}
\arguments{
\item{x}{object of class optimsens}

\item{...}{other stuff (for S3 consistency)}
}
\description{
\code{print} method of object class \code{\link{optimsens}}
}
