% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.policy}
\alias{print.policy}
\title{\code{print} method of object class \code{\link{policy}}}
\usage{
\method{print}{policy}(x, ...)
}
\arguments{
\item{x}{an object class policy}

\item{...}{other stuff}
}
\description{
\code{print} method of object class \code{\link{policy}}
}
