% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sens}
\alias{print.sens}
\title{\code{print} method of object class \code{sens} (parent of
\code{\link{optimsens}}, and \code{\link{gridsens}})}
\usage{
\method{print}{sens}(x, ...)
}
\arguments{
\item{x}{object of class sens}

\item{...}{other stuff (for S3 consistency)}
}
\description{
\code{print} method of object class \code{sens} (parent of
\code{\link{optimsens}}, and \code{\link{gridsens}})
}
