% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.undi}
\alias{print.undi}
\title{\code{print} method of object class \code{\link{undi}}}
\usage{
\method{print}{undi}(x, ...)
}
\arguments{
\item{x}{an object class undi}

\item{...}{other stuff}
}
\description{
\code{print} method of object class \code{\link{undi}}
}
