% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitize.R
\name{sensitize.policy}
\alias{sensitize.policy}
\title{Compute the sensitivity-adjusted estimates of predicted outcome given
treatment/control for a policy object}
\usage{
\method{sensitize}{policy}(obj, q, dp, d0, d1, compare = NULL,
  ptreat = NULL, resp_ctl = NULL, resp_trt = NULL, ...)
}
\arguments{
\item{obj}{data to sensitize}

\item{q}{p(u = 1 | x) (see Details)}

\item{dp}{change in log-odds of treat = 1 if u = 1 (see Details)}

\item{d0}{change in log-odds of response = 1 if treat = 0 and u = 1 (see
Details)}

\item{d1}{change in log-odds of response = 1 of treat = 1 and u = 1 (see
Details)}

\item{compare}{(Optional) character vector of groups to compare; the data
will be filtered such that only specified groups are compared, and the
\code{grouping} variable will be refactored such that the levels preserve
the specified order, e.g., \code{compare = c("white", "black")} will make
\code{"white"} the base group}

\item{ptreat}{(Optional) default value for probability of treatment; if
provided, it will override fitted values in \code{pol$data}}

\item{resp_ctl}{(Optional)}

\item{resp_trt}{(Optional) default value for probability of response = 1
given each treatment regime (\code{ctl}, \code{trt}); useful for cases
where outcome under certain treatment regimes is deterministic (e.g.,
probability of finding illegal weapon if NOT frisked is 0); if provided, it
will override fitted values in \code{pol$data}}

\item{...}{additional arguments are ignored}
}
\value{
a new \code{sensitive_policy} object, which inherits \code{policy},
  with the data element updated according to sensitivity parameters; all
  other aspects of the original \code{policy} object are preserved.
}
\description{
Compute the sensitivity-adjusted estimates of predicted outcome given
treatment/control for a policy object
}
\details{
All sensitivity parameters (\code{q, dp, d0, d1}) can be provided in
  one of three formats, determined by the \code{length} of each argument:
  \describe{ \item{if \code{length(arg) = 1}}{single value applied to all
  observations (rows)} \item{if \code{length(arg) = }number of levels in
  grouping variable}{each parameter setting applied to corresponding level in
  group} \item{if \code{length(arg) = nrow(pol$data)}}{each parameter applied
  to corresponding rows}} Note that if \code{compare} is specified, the
  number of grouping levels is effectively the length of \code{compare}
}
