% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.optimsens}
\alias{summary.optimsens}
\title{Summary of optimsens object}
\usage{
\method{summary}{optimsens}(object, ...)
}
\arguments{
\item{object}{optimsens object}

\item{...}{other stuff}
}
\value{
summary data frame with min/max values of each group term
}
\description{
Summary of optimsens object
}
