% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.sens}
\alias{summary.sens}
\title{Summary of sens object}
\usage{
\method{summary}{sens}(object, ...)
}
\arguments{
\item{object}{\code{sens} object}

\item{...}{other stuff}
}
\value{
summary data frame with min/max values of each group term
}
\description{
Summary of sens object
}
