% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{synthesize}
\alias{synthesize}
\title{Generate synthetic sample of a policy}
\usage{
synthesize(pol, seed = round(stats::runif(1) * 10000))
}
\arguments{
\item{pol}{policy object}

\item{seed}{(Optional) random seed to set before generating synthetic
observations}
}
\value{
new policy object with the treatment and outcome columns in
  \code{$data} updated with synthetic samples based on estimated values of
  \code{ptrt__}, \code{resp_ctl__}, and \code{resp_trt__}
}
\description{
Generate synthetic sample of a policy
}
