% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitize.R
\name{undisens}
\alias{undisens}
\title{Sensitivity to unobserved confounders with specified parameters}
\usage{
undisens(u, q, dp, d0, d1, ptreat = NULL, resp_ctl = NULL,
  resp_trt = NULL)
}
\arguments{
\item{u}{object of class \code{undi}}

\item{q}{p(u = 1 | x) (see Details)}

\item{dp}{change in log-odds of treat = 1 if u = 1 (see Details)}

\item{d0}{change in log-odds of response = 1 if treat = 0 and u = 1 (see
Details)}

\item{d1}{change in log-odds of response = 1 of treat = 1 and u = 1 (see
Details)}

\item{ptreat}{(Optional) default value for probability of treatment; if
provided, it will override fitted values in \code{u$data}}

\item{resp_ctl}{(Optional)}

\item{resp_trt}{(Optional) default value for probability of response = 1
given each treatment regime (\code{ctl}, \code{trt}); useful for cases
where outcome under certain treatment regimes is deterministic (e.g.,
probability of finding illegal weapon if NOT frisked is 0); if provided,
it will override fitted values in \code{u$data}}
}
\value{
\code{tidy} dataframe of second-stage model coefficients after
  applying sensitivity parameters, with a nested column of sensitivity params
}
\description{
Sensitivity to unobserved confounders with specified parameters
}
\details{
All sensitivity parameters (\code{q, dp, d0, d1}) can be provided in
  one of three formats, determined by the \code{length} of each argument:
  \describe{ \item{if \code{length(arg) = 1}}{single value applied to all
  observations (rows)} \item{if \code{length(arg) = }number of levels in
  grouping variable}{each parameter setting applied to corresponding level in
  group} \item{if \code{length(arg) = nrow(u$data)}}{each parameter applied
  to corresponding rows}}
}
