% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{compute_rad}
\alias{compute_rad}
\title{Run test for unjustified disparate impact}
\usage{
compute_rad(pol, controls = NULL, base_group = NULL,
  minority_groups = NULL)
}
\arguments{
\item{pol}{object of class policy}

\item{controls}{character vector of additional controls to consider in the
second-stage model}

\item{base_group}{(Optional) single group that acts as the pivot/base; by
default, if the grouping variable is a factor, set to the first level,
otherwise set to the first of sorted unique values}

\item{minority_groups}{(Optional) groups to compare to the base group; by
default, set to every unique value other than the base group}
}
\value{
tidy data frame of rad coefficients
}
\description{
Run test for unjustified disparate impact
}
