% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compute_rad}
\alias{compute_rad}
\title{Compute risk-adjusted disparate impact estimate for a policy}
\usage{
compute_rad(pol, controls = NULL, base_group = NULL,
  minority_groups = NULL, fit_fn = c("logit", "gam"), down_sample = 1,
  seed = round(stats::runif(1) * 10000), ...)
}
\arguments{
\item{pol}{object of class policy}

\item{controls}{character vector of additional controls to consider in the
second-stage model}

\item{base_group}{(Optional) single group that acts as the pivot/base; by
default, if the grouping variable is a factor, set to the first level,
otherwise set to the first of sorted unique values}

\item{minority_groups}{(Optional) groups to compare to the base group; by
default, set to every unique value other than the base group}

\item{fit_fn}{string indicating the fitting proceedure used.}

\item{down_sample}{(Optional) proportion (between 0 and 1) or number (greater
than 1) of rows to sample, if down sampling the (test) data; default is 1
(i.e., use all data)}

\item{seed}{random seed to set}

\item{...}{additional arguments to pass to \code{fit} function from
\code{\link{di_model}} for fine-tuning}
}
\value{
tidy data frame with columns \item{term}{the group members considered
  minority} \item{estimate}{the treatment ratio of each minority group over
  base group} \item{std.error}{TO BE IMPLEMENTED! FOR NOW, ALL ZERO}
  \item{ptrt_base/minor}{estimated average treatment probability for base/minority
  groups} \item{controls}{features controlled for}
}
\description{
Compute risk-adjusted disparate impact estimate for a policy
}
