% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_sgd}
\alias{fit_sgd}
\title{Wrapper for fitting \code{\link[sgd]{sgd}} models with standardized design
matrix}
\usage{
fit_sgd(f, d, ...)
}
\arguments{
\item{f}{formula}

\item{d}{data object}

\item{...}{additional arguments for \code{\link[sgd]{sgd}}}
}
\value{
an \code{sgd} object, recast as \code{class} \code{sgd.scaled}, with
  additional attributes from \code{\link[base]{scale}}
}
\description{
Wrapper for fitting \code{\link[sgd]{sgd}} models with standardized design
matrix
}
\details{
\code{sgd} models fit with scaling \strong{must} use accompanying
  \code{\link{predict_sgd}} function to generate properly scaled predictions
}
