% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optimsens}
\alias{optimsens}
\title{Find parameter values that min/maximize undi results}
\usage{
optimsens(r, range_q = c(0, 1), range_dp = c(0, log(2)), range_d0 = c(0,
  log(2)), range_d1 = c(0, log(2)), base_group = NULL,
  minority_groups = NULL, debug = FALSE)
}
\arguments{
\item{r}{object of class \code{undi}}

\item{range_q}{2D vector specifying min/max value of p(u = 1 | x)}

\item{range_dp}{2D vector specifying min/max value of change in log-odds of
treat = 1 if u = 1}

\item{range_d0}{2D vector specifying min/max value of change in log-odds of
response = 1 if treat = 0 and u = 1}

\item{range_d1}{2D vector specifying min/max value of change in log-odds of
response = 1 of treat = 1 and u = 1}

\item{base_group}{(Optional) single group that acts as the pivot/base; by
default, if the grouping variable is a factor, set to the first level,
otherwise set to the first of sorted unique values}

\item{minority_groups}{(Optional) groups to compare to the base group; by
default, set to every unique value other than the base group}

\item{debug}{logical flag, if TRUE, returns a list of results and the
expanded data frame used to fit model}
}
\value{
\code{tidy} dataframe of second-stage model coefficients after
  applying sensitivity parameters, with a nested column of sensitivity params
}
\description{
Within specified range of sensitivity parameters, find the ones that achieve
minimum/maximum undi results
}
\details{
All sensitivity parameters (\code{q, dp, d0, d1}) can be provided in
  one of three formats, determined by the \code{length} of each argument:
  \describe{ \item{if \code{length(arg) = 1}}{single value applied to all
  observations (rows)} \item{if \code{length(arg) = }number of levels in
  grouping variable}{each parameter setting applied to corresponding level in
  group} \item{if \code{length(arg) = nrow(u$data)}}{each parameter applied
  to corresponding rows}} Note that if \code{compare} is specified, the
  number of grouping levels is effectively the length of \code{compare}
}
