% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.policy}
\alias{plot.policy}
\title{Plot risk-treatment and calibration plots for a \code{policy}}
\usage{
\method{plot}{policy}(x, nbins = 10, method = "glm",
  method.args = list(family = "quasibinomial"), ...)
}
\arguments{
\item{x}{object of class \code{policy}}

\item{nbins}{number of bins to use (e.g., for calibration plots)}

\item{method}{passed to \code{\link[ggplot2]{geom_smooth}}}

\item{method.args}{passed to \code{\link[ggplot2]{geom_smooth}}}

\item{...}{ignored; included for S3 generic/method consistency}
}
\value{
a named list of ggplot object
}
\description{
Plot risk-treatment and calibration plots for a \code{policy}
}
