% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{policy}
\alias{policy}
\title{Create policy object}
\usage{
policy(formula, data, outcome, controls = NULL, train = 0.5, fit1 = NULL,
  pred1 = NULL, fit2 = NULL, fit_ptreat = NULL, pred_ptreat = NULL,
  risk = "resp_trt", ptreat = NULL, resp_ctl = NULL, resp_trt = NULL,
  calibrate = FALSE, seed = 1234, ...)
}
\arguments{
\item{formula}{a formula in the form of \code{treatment ~ grouping_variable +
other predictors} where the LHS is the treatment column, first element on
the RHS is the grouping variable (e.g., Race), and the remainder of the RHS
specifies the predictors for first-stage model}

\item{data}{data frame to use; must include all the columns specified in
\code{formula} and given in the \code{outcome} parameter}

\item{outcome}{name of outcome column in data}

\item{controls}{character vector of additional controls to consider in the
second-stage model}

\item{train}{either (1) a value between 0 and 1 representing the proportion
of data to use in training, (2) the name of a column of characters "train"
and "test" within \code{data} to use in splitting the data, or (3) a
logical vector of equal length as \code{nrow(data)} used to index training
data}

\item{fit1}{a function of the form f(formula, data, ...) used for fitting the
first-stage model; using \code{gbm} by default}

\item{pred1}{a function of the form f(model, data, formula) used for
generating predictions from the first-stage model; the formula argument can
be ignored within the function body, but the function should still accept
it; some prediction functions (e.g., glmnet) require the original formula;
predictions should be on probability scale, while "risk" will always be on
logit scale}

\item{fit2}{a function of the form f(formula, data, weights = NULL) used for
fitting the second-stage model; using \code{glm} with \code{family =
quasibinomial} by default; the \code{weights} argument is only used for
sensitivity and *must* be initialized to \code{NULL} (or the equivalent of
non-weighted fitting)}

\item{fit_ptreat}{a function of the form f(formula, data, ...) used for
fitting propensity (probability of treatment) models. If not specified,
\code{fit1} is used by default, with the provided \code{formula} argument.}

\item{pred_ptreat}{a function of the form f(model, data, formula) used for
generating propensity predictions. If not specified, \code{pred1} is used
by default.}

\item{risk}{One of \code{"resp_ctl"} or \code{"resp_trt"}, indicating which
treatment regime should be used as the risk score (default:
\code{resp_trt})}

\item{ptreat}{(Optional) default value for probability of treatment; if
provided, it will override \code{fit_ptreat} and \code{pred_ptreat}}

\item{resp_ctl}{(Optional)}

\item{resp_trt}{(Optional) default value for probability of response = 1
given each treatment regime (\code{ctl}, \code{trt}); useful for cases
where outcome under certain treatment regimes is deterministic (e.g.,
probability of finding illegal weapon if NOT frisked is 0)}

\item{calibrate}{whether or not to use platt scaling to calibrate predictions}

\item{seed}{random seed to use}

\item{...}{additional arguments passed to first-stage model fitting function,
\code{fit1} and \code{fit_ptreat}}
}
\value{
policy object \item{data}{original data frame, augmented with columns
  \code{fold__}, \code{ptrt__}, \code{resp_ctl__}, \code{resp_trt__}, and
  \code{risk__}, which contain train/test fold indicators for the first
  stage, treatment propesinty, first stage model probability predictions
  under control and treatment, and appropriate risk measure on logit-scale,
  respectively} \item{risk_col}{either "resp_ctl" or "resp_trt", indicating
  which was used for "risk"}\item{m1_*}{fitted first stage model for response
  give ctl/trt} \item{treatment}{name of column from \code{data} used as
  treatment indicator} \item{outcome}{name of column from \code{data} used as
  outcome indicator} \item{grouping}{name of column from \code{data} used as
  grouping variable} \item{features}{additional features used in first stage
  model} \item{controls}{legitimate controls used in second stage model}
  \item{fit1}{function used to fit first stage model} \item{pred1}{function
  used to generate predictions from first stage model} \item{fit2}{function
  used to fit second stage model} \item{fit_ptreat}{function used to fit
  model for treatment propensity} \item{pred_ptreat}{function used to
  generate predictions for treatment propensity}
}
\description{
Create policy object
}
