% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.optimsens}
\alias{summary.optimsens}
\title{Summary of optimsens object}
\usage{
\method{summary}{optimsens}(x, ...)
}
\arguments{
\item{x}{optimsens object}

\item{...}{other stuff}
}
\description{
Summary of optimsens object
}
