% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undi.R
\docType{package}
\name{undi}
\alias{undi}
\alias{undi-package}
\title{undi: Testing for unjustified disparate impact}
\description{
The \code{undi} package implements a test for unjustified disparate impact in
policy decisions.
}
